qpnp-smbcharger-mmi: Hide charge counter
Querying the charge counter during deep sleep takes an excessively long
amount of time. BatteryService looks up this prop during deep sleep,
resulting in a very slow wakeup on some devices.

Add a filter to hide the charge counter, so BatteryService does not
attempt to read the prop. It is unknown as to whether this is a
firmware or a hardware issue.

Signed-off-by: Rohit Vashisht <rohitv88cf@gmail.com>